refactor(commands): Extract REPO_NAME_PATTERN and use ephemeral replies (#37, #38)
- Extract REPO_NAME_PATTERN from src/index.ts to src/utils/validation.ts to avoid circular import concerns
- Re-export from index.ts for backward compatibility
- Update command handlers and tests to import directly from validation module
- Use ephemeral replies for validation error messages in /pr, /issue, and /status commands
