+++ src/commands/status.ts
@@ -11,7 +11,7 @@ import {
   Colors,
 } from 'discord.js';
 import { safeErrorMessage } from '../utils/errors.js';
-import { REPO_NAME_PATTERN } from '../index.js';
+import { REPO_NAME_PATTERN } from '../utils/validation.js';
 
 interface GitHubPr {
   number: number;
@@ -40,7 +40,7 @@ export async function handleStatusCommand(
   repo: string
 ): Promise<void> {
   if (!REPO_NAME_PATTERN.test(repo)) {
-    await interaction.reply('Invalid repository format');
+    await interaction.reply({ content: 'Invalid repository format', ephemeral: true });
     return;
   }
