+++ src/utils/validation.ts
@@ -1,0 +1,2 @@
+/** Validates `owner/repo` format for GitHub repository names. */
+export const REPO_NAME_PATTERN = /^[a-zA-Z0-9._-]+\/[a-zA-Z0-9._-]+$/;
